Cleaned up code
- Moved files around
- Removed Nav for Header
- Made contributors its own page
- Fixed importing of components to fixed path
