+++ docs/.vuepress/theme/components/Footer.vue
@@ -1,35 +1,10 @@
 <template>
     <footer style="background: linear-gradient(90deg, #064789 0%, #427aa1 100%)">
-      <div class="container mx-auto px-6 pt-10 pb-6">
-        <div class="flex flex-wrap">
-          <div class="w-full md:w-1/2 text-center md:text-right"></div>
-          
-          <div class="w-full md:w-1/4 text-right md:text-left">
-            <h5 class="uppercase mb-6 font-bold text-white border-teal-500 border-b-2 mr-40">Links</h5>
-            <ul class="mb-4">
-              <li class="mt-2">
-                <a href="https://discord.gg/Huhe48c" class="hover:underline text-white hover:text-teal-500">Help</a>
-              </li>
-              <li class="mt-2">
-                <a href="https://github.com/jamiepine/pulse/issues" class="hover:underline text-white hover:text-teal-500">Support</a>
-              </li>
-            </ul>
-          </div>
-          <div class="w-full md:w-1/4 text-right  md:text-left">
-            <h5 class="uppercase mb-6 font-bold text-white border-teal-500 border-b-2 mr-40">Social</h5>
-            <ul class="mb-4">
-              <li class="mt-2">
-                <a href="https://twitch.tv/jamiepinelive" class="hover:underline text-white hover:text-teal-500">Twitch</a>
-              </li>
-              <li class="mt-2">
-                <a href="https://twitter.com/jamiepine" class="hover:underline text-white hover:text-teal-500">Twitter</a>
-              </li>
-            </ul>
-          </div>
-        </div>
-      </div>
       <div class="container mx-auto flex items-center justify-center p-2 pb-4">
           <div class="text-gray-400 text-center">©2019 James Pine. All rights reserved.</div>
+      </div>
+      <div>
+        <!-- FONT AWESOME SOCIAL ICONS HERE -->
       </div>
       <div class="container mx-auto flex items-center justify-center pb-5">
         <div class="flex">
