+++ docs/.vuepress/theme/index.js
@@ -1,22 +1,9 @@
 module.exports = {
+    extend: '@vuepress/theme-default',
     plugins: [
         ['@vuepress/search', {
             searchMaxSuggestions: 10
         }],
         ['@vuepress/active-header-links'],
     ],
-    markdown: {
-        /**
-         * Add class named to html generated by Markdown
-         */
-        lineNumbers: true,
-        extendMarkdown: md => {
-          md.set({ html: true })
-          md.use(require('markdown-it-highlightjs'))
-          md.use(require('markdown-it-container'))
-          md.use(require('markdown-it-plantuml'))
-          md.use(require('markdown-it-admonition'))
-        },
-        anchor: { permalink: false }
-    },
 }
